+++ Runtime/Extras/EyeAnimationHandler.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections;
+﻿using System.Threading;
+using System.Threading.Tasks;
 using UnityEngine;
 
 namespace ReadyPlayerMe.AvatarLoader
@@ -18,7 +19,9 @@ namespace ReadyPlayerMe.AvatarLoader
         private const float HALFBODY_OFFSET_X = 90;
         private const float HALFBODY_OFFSET_Z = 180;
         private const string MISSING_EYE_BONES_MESSAGE = "Eye bones are required for EyeAnimationHandler.cs but they were not found on loaded Avatar! Eye rotation animations will not be applied";
-        private const string MISSING_MORPH_TARGETS_MESSAGE = "The 'eyeBlinkLeft' & 'eyeBlinkRight' morph targets are required for EyeAnimationHandler.cs but they were not found on Avatar mesh! Use an AvatarConfig to specify the morph targets to be included on loaded avatars.";
+        private const string MISSING_MORPH_TARGETS_MESSAGE =
+            "The 'eyeBlinkLeft' & 'eyeBlinkRight' morph targets are required for EyeAnimationHandler.cs but they were not found on Avatar mesh! Use an AvatarConfig to specify the morph targets to be included on loaded avatars.";
+        private const string NO_EYE_BONES_FOUND_IN_AVATAR_SKELETON = "No eye bones found in Avatar skeleton!";
 
         [SerializeField, Range(0, 1), Tooltip("Effects the duration of the avatar blink animation in seconds.")]
         private float blinkDuration = 0.1f;
@@ -27,7 +30,7 @@ namespace ReadyPlayerMe.AvatarLoader
         private float blinkInterval = 3f;
 
         private WaitForSeconds blinkDelay;
-        private Coroutine blinkCoroutine;
+        private Task blinkCoroutine;
 
         private SkinnedMeshRenderer headMesh;
         private int eyeBlinkLeftBlendShapeIndex = -1;
@@ -41,6 +44,7 @@ namespace ReadyPlayerMe.AvatarLoader
         private bool hasEyeBones;
         private bool CanAnimate => hasBlinkBlendShapes || hasEyeBones;
 
+        private CancellationTokenSource ctxSource;
         public float BlinkDuration
         {
             get => blinkDuration;
@@ -80,7 +84,7 @@ namespace ReadyPlayerMe.AvatarLoader
             headMesh = gameObject.GetMeshRenderer(MeshType.HeadMesh);
             hasBlinkBlendShapes = HasBlinkBlendshapes();
             ValidateSkeleton();
-            
+
             if (!hasBlinkBlendShapes)
             {
                 Debug.LogWarning(MISSING_MORPH_TARGETS_MESSAGE);
@@ -89,21 +93,21 @@ namespace ReadyPlayerMe.AvatarLoader
             {
                 Debug.LogWarning(MISSING_EYE_BONES_MESSAGE);
             }
-            
+
             if (!CanAnimate)
             {
                 Reset();
                 enabled = false;
             }
         }
-        
+
         private bool HasBlinkBlendshapes()
         {
             eyeBlinkLeftBlendShapeIndex = headMesh.sharedMesh.GetBlendShapeIndex(EYE_BLINK_LEFT_BLEND_SHAPE_NAME);
             eyeBlinkRightBlendShapeIndex = headMesh.sharedMesh.GetBlendShapeIndex(EYE_BLINK_RIGHT_BLEND_SHAPE_NAME);
             return eyeBlinkLeftBlendShapeIndex > -1 && eyeBlinkRightBlendShapeIndex > -1;
         }
-        
+
         private void ValidateSkeleton()
         {
             isFullBody = AvatarBoneHelper.IsFullBodySkeleton(transform);
@@ -112,16 +116,16 @@ namespace ReadyPlayerMe.AvatarLoader
             hasEyeBones = leftEyeBone != null && rightEyeBone != null;
             if (!hasEyeBones)
             {
-                Debug.LogWarning("No eye bones found in Avatar skeleton!");
+                Debug.LogWarning(NO_EYE_BONES_FOUND_IN_AVATAR_SKELETON);
             }
         }
 
         private void Reset()
         {
             CancelInvoke();
-            blinkCoroutine?.Stop();
+            ctxSource?.Cancel();
         }
-        
+
         private void OnEnable()
         {
             if (CanAnimate)
@@ -143,7 +147,7 @@ namespace ReadyPlayerMe.AvatarLoader
         /// <summary>
         /// Rotates the eyes and assigns the blink coroutine. Called in the Initialize method.
         /// </summary>
-        private void AnimateEyes()
+        private async void AnimateEyes()
         {
             if (hasEyeBones)
             {
@@ -152,7 +156,8 @@ namespace ReadyPlayerMe.AvatarLoader
 
             if (hasBlinkBlendShapes)
             {
-                blinkCoroutine = BlinkEyes().Run();
+                ctxSource = new CancellationTokenSource();
+                await BlinkEyes(ctxSource.Token);
             }
         }
 
@@ -168,7 +173,7 @@ namespace ReadyPlayerMe.AvatarLoader
         {
             float vertical = Random.Range(-VERTICAL_MARGIN, VERTICAL_MARGIN);
             float horizontal = Random.Range(-HORIZONTAL_MARGIN, HORIZONTAL_MARGIN);
-            
+
             Quaternion rotation = isFullBody
                 ? Quaternion.Euler(horizontal, vertical, 0)
                 : Quaternion.Euler(horizontal - HALFBODY_OFFSET_X, 0, vertical + HALFBODY_OFFSET_Z);
@@ -178,13 +183,20 @@ namespace ReadyPlayerMe.AvatarLoader
         /// <summary>
         /// A coroutine that manipulates BlendShapes to open and close the eyes.
         /// </summary>
-        private IEnumerator BlinkEyes()
+        /// <param name="ctx"></param>
+        private async Task BlinkEyes(CancellationToken ctx)
         {
             headMesh.SetBlendShapeWeight(eyeBlinkLeftBlendShapeIndex, EYE_BLINK_MULTIPLIER);
             headMesh.SetBlendShapeWeight(eyeBlinkRightBlendShapeIndex, EYE_BLINK_MULTIPLIER);
 
-            yield return blinkDelay;
-
+            var startTime = Time.time;
+            while (Time.time - startTime < blinkDuration && !ctx.IsCancellationRequested)
+            {
+                await Task.Yield();
+            }
+            
+            if(ctx.IsCancellationRequested) return;
+            
             headMesh.SetBlendShapeWeight(eyeBlinkLeftBlendShapeIndex, 0);
             headMesh.SetBlendShapeWeight(eyeBlinkRightBlendShapeIndex, 0);
         }