+++ Runtime/Operations/AvatarDownloader.cs
@@ -13,6 +13,8 @@ namespace ReadyPlayerMe.AvatarLoader
     public class AvatarDownloader : IOperation<AvatarContext>
     {
         private const string TAG = nameof(AvatarDownloader);
+        private const string DOWNLOADING_AVATAR_INTO_MEMORY = "Downloading avatar into memory.";
+        private const string LOADING_MODEL_FROM_CACHE = "Loading model from cache.";
 
         /// If true the avatar will download into memory instead of a local file.
         private readonly bool downloadInMemory;
@@ -54,16 +56,16 @@ namespace ReadyPlayerMe.AvatarLoader
             }
 
             if ((!context.IsUpdateRequired || Application.internetReachability == NetworkReachability.NotReachable)
-                    && File.Exists(context.AvatarUri.LocalModelPath))
+                && File.Exists(context.AvatarUri.LocalModelPath))
             {
-                SDKLogger.Log(TAG, "Loading model from cache.");
+                SDKLogger.Log(TAG, LOADING_MODEL_FROM_CACHE);
                 context.Bytes = File.ReadAllBytes(context.AvatarUri.LocalModelPath);
                 return context;
             }
 
             if (context.IsUpdateRequired)
             {
-                AvatarCache.ClearAvatar(context.AvatarUri.Guid, context.SaveInProjectFolder);
+                AvatarCache.DeleteAvatarModel(context.AvatarUri.Guid, context.SaveInProjectFolder);
             }
 
             if (downloadInMemory)
@@ -95,14 +97,14 @@ namespace ReadyPlayerMe.AvatarLoader
                 SDKLogger.Log(TAG, $"Download URL with parameters: {url}");
             }
 
-            SDKLogger.Log(TAG, "Downloading avatar into memory.");
+            SDKLogger.Log(TAG, DOWNLOADING_AVATAR_INTO_MEMORY);
 
             var dispatcher = new WebRequestDispatcher();
             dispatcher.ProgressChanged = ProgressChanged;
 
             try
             {
-                var response = await dispatcher.DownloadIntoMemory(url, token, Timeout);
+                Response response = await dispatcher.DownloadIntoMemory(url, token, Timeout);
                 return response.Data;
             }
             catch (Exception exception)
@@ -138,7 +140,7 @@ namespace ReadyPlayerMe.AvatarLoader
 
             try
             {
-                var response = await dispatcher.DownloadIntoFile(url, path, token, Timeout);
+                ResponseFile response = await dispatcher.DownloadIntoFile(url, path, token, Timeout);
                 return response.Data;
             }
             catch (Exception exception)