+++ Runtime/Operations/RenderParameterProcessor.cs
@@ -1,5 +1,3 @@
-
-
 namespace ReadyPlayerMe.AvatarLoader
 {
     public static class RenderParameterProcessor