+++ Runtime/Utils/AvatarBoneHelper.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 /// This static class contains useful helper functions used check and access
 /// bones of a loaded avatar.
 /// </summary>
-public static class AvatarBoneHelper 
+public static class AvatarBoneHelper
 {
     private const string ARMATURE_HIPS_LEFT_UP_LEG_BONE_NAME = "Armature/Hips/LeftUpLeg";
     private const string HALF_BODY_LEFT_EYE_BONE_NAME = "Armature/Hips/Spine/Neck/Head/LeftEye";
     private const string FULL_BODY_LEFT_EYE_BONE_NAME = "Armature/Hips/Spine/Spine1/Spine2/Neck/Head/LeftEye";
     private const string HALF_BODY_RIGHT_EYE_BONE_NAME = "Armature/Hips/Spine/Neck/Head/RightEye";
     private const string FULL_BODY_RIGHT_EYE_BONE_NAME = "Armature/Hips/Spine/Spine1/Spine2/Neck/Head/RightEye";
-    
+
     /// <summary>
     /// This is a legacy function that can be used to check if an
     /// avatar is fullbody and the object does not have an <seealso cref="AvatarData"/> component
@@ -22,8 +22,8 @@ public static class AvatarBoneHelper
     {
         return avatarRoot.Find(ARMATURE_HIPS_LEFT_UP_LEG_BONE_NAME);
     }
-    
-    
+
+
     /// <summary>
     /// Searches the avatar transform hierarchy for the left eye bone
     /// </summary>
@@ -34,7 +34,7 @@ public static class AvatarBoneHelper
     {
         return avatarRoot.Find(isFullBody ? FULL_BODY_LEFT_EYE_BONE_NAME : HALF_BODY_LEFT_EYE_BONE_NAME);
     }
-    
+
     /// <summary>
     /// Searches the avatar transform hierarchy for the right eye bone
     /// </summary>