+++ Tests/Editor/AvatarDownloadTests.cs
@@ -37,9 +37,12 @@ namespace ReadyPlayerMe.AvatarLoader.Tests
         [Test]
         public async Task Download_Avatar_Into_Memory()
         {
-            Cleanup();
+            if (File.Exists(TestAvatarData.DefaultAvatarUri.LocalModelPath))
+            {
+                File.Delete(TestAvatarData.DefaultAvatarUri.LocalModelPath);
+            }
             byte[] bytes;
-            
+
             var avatarDownloader = new AvatarDownloader();
             try
             {