+++ Tests/Editor/AvatarLoaderTests.cs
@@ -1,23 +1,32 @@
+using System;
 using System.Collections;
+using System.Threading.Tasks;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
+using Object = UnityEngine.Object;
 
 namespace ReadyPlayerMe.AvatarLoader.Tests
 {
     public class AvatarLoaderTests
     {
+        private GameObject avatar;
+        
         [TearDown]
         public void Cleanup()
         {
             TestUtils.DeleteAvatarDirectoryIfExists(TestAvatarData.DefaultAvatarUri.Guid, true);
             TestUtils.DeleteAvatarDirectoryIfExists(TestUtils.TEST_WRONG_GUID, true);
+
+            if (avatar != null)
+            {
+                Object.DestroyImmediate(avatar);
+            }
         }
 
         [UnityTest]
         public IEnumerator AvatarLoader_Complete_Load()
         {
-            GameObject avatar = null;
             var avatarUrl = string.Empty;
             var failureType = FailureType.None;
 
@@ -36,6 +45,31 @@ namespace ReadyPlayerMe.AvatarLoader.Tests
             Assert.AreEqual(FailureType.None, failureType);
             Assert.IsNotNull(avatar);
             Assert.IsNotNull(avatar.GetComponent<AvatarData>());
+        }
+
+        [Test]
+        public async Task AvatarLoader_Complete_Load_Async()
+        {
+            var loader = new AvatarObjectLoader();
+            EventArgs args = await loader.LoadAvatarAsync(TestAvatarData.DefaultAvatarUri.ModelUrl);
+
+            Assert.True(args != null);
+
+            if (args.GetType() == typeof(FailureEventArgs))
+            {
+                Assert.Fail(((FailureEventArgs) args).Type.ToString());
+            }
+            else if (args.GetType() == typeof(CompletionEventArgs))
+            {
+                var completedEventArgs = (CompletionEventArgs) args;
+                Assert.AreEqual(TestAvatarData.DefaultAvatarUri.ModelUrl, completedEventArgs.Url);
+                Assert.IsNotNull(completedEventArgs.Avatar);
+                avatar = completedEventArgs.Avatar;
+            }
+            else
+            {
+                Assert.Fail("Unknown event args type");
+            }
         }
 
         [UnityTest]
@@ -91,15 +125,14 @@ namespace ReadyPlayerMe.AvatarLoader.Tests
             AvatarLoaderSettings settings = AvatarLoaderSettings.LoadSettings();
             settings.AvatarCachingEnabled = true;
 
-            GameObject avatarA = null;
             var failureType = FailureType.None;
 
             var loader = new AvatarObjectLoader();
-            loader.OnCompleted += (_, args) => avatarA = args.Avatar;
+            loader.OnCompleted += (_, args) => avatar = args.Avatar;
             loader.OnFailed += (_, args) => failureType = args.Type;
             loader.LoadAvatar(TestAvatarData.DefaultAvatarUri.ModelUrl);
 
-            yield return new WaitUntil(() => avatarA != null || failureType != FailureType.None);
+            yield return new WaitUntil(() => avatar != null || failureType != FailureType.None);
 
             Assert.AreEqual(FailureType.None, failureType);
             Assert.AreEqual(false, AvatarCache.IsCacheEmpty());
@@ -113,7 +146,6 @@ namespace ReadyPlayerMe.AvatarLoader.Tests
         [UnityTest]
         public IEnumerator AvatarLoader_Cancel_Loading()
         {
-            GameObject avatar = null;
             var failureType = FailureType.None;
             var loader = new AvatarObjectLoader();