+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.0] - UNRELEASED
+
+### Added
+- API for async avatar loading [#76](https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader/pull/76)
+- added setup guide window [#85](https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader/pull/85)
+
+### Updated
+- updates for core module refactor [#78](https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader/pull/78)
+- general cleanup and refactor [#79](https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader/pull/79)
+- refactor of multiple avatar renderer example [#83](https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader/pull/83)
+- refactor of avatar cache [#86](https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader/pull/86)
+- documentation UI improvements [#87](https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader/pull/87)
+
+### Fixed
+- alignment issues and spacing in Core settings window [#88](https://github.com/readyplayerme/rpm-unity-sdk-avatar-loader/pull/88)
+
 ## [1.2.0] - 2023.04.18
 
 ### Added